2000-02-17  bertrand  <Bertrand.Guiheneuf@aful.org>
	* camel/camel-formatter.c (handle_text_plain):
	revamped so that it uses the output stream
	of the data wrapper
	(handle_text_html): ditto.


	* camel/camel-simple-data-wrapper.h:
	* camel/camel-simple-data-wrapper.c (camel_simple_data_wrapper_new):
	use (void) instead of ().
	(_get_output_stream): simple implementation.

A lot of small fixes so that the new parser scheme
works properly. Simple implementation of the stream
though.

Changed vette-formatter files so that they work with the
new scheme.

The new parser is now in a usable state. Still needs some work but the
infrastructure is here. /me is happy.


git-svn-id: svn+ssh://svn.gnome.org/svn/evolution-data-server/trunk@273 d0434b6f-c725-0410-8785-fd8a057797ef